[mod/#127] Navi 구조 수정
